Move lyric time-tag display area into single layer.
